♻️ Refactor : 리뷰 상세내용 추가

+++ src/main/java/umc/umcspring/web/dto/ReviewRequestDTO.java
@@ -13,7 +13,7 @@ public class ReviewRequestDTO {
     public static class AddReviewDTO{
         String body;
         @NotNull
-        Integer score;
+        Float score;
         @ExistMember
         Long memberId;
         @ExistMemberMission
